add dependencies to perfDriver
