+++ pom.xml
@@ -19,6 +19,7 @@
         <module>generator</module>
         <module>migration</module>
         <module>perfClient</module>
+        <module>perfDriver</module>
 
     </modules>
